+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -15,6 +15,7 @@ mod read_bytes;
 mod read_text;
 mod rm;
 mod rmdir;
+mod set_mode;
 mod temp_file;
 mod touch;
 mod write_bytes;
@@ -46,6 +47,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(read_text::create(&package))?;
     commands.set(rm::create(&package))?;
     commands.set(rmdir::create(&package))?;
+    commands.set(set_mode::create(&package))?;
     commands.set(temp_file::create(&package))?;
     commands.set(touch::create(&package))?;
     commands.set(write_bytes::create(&package))?;