+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -4,6 +4,7 @@ mod canonical;
 mod cp;
 mod dirname;
 mod exists;
+mod glob_array;
 mod is_directory;
 mod is_file;
 mod is_readonly;
@@ -36,6 +37,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(cp::create(&package))?;
     commands.set(dirname::create(&package))?;
     commands.set(exists::create(&package))?;
+    commands.set(glob_array::create(&package))?;
     commands.set(is_directory::create(&package))?;
     commands.set(is_file::create(&package))?;
     commands.set(is_readonly::create(&package))?;