+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### v0.2.2
 
-* New is_directory command
-* New is_file command
+* New is_path_exists command.
+* New is_directory command.
+* New is_file command.
 
 ### v0.2.1 (2020-02-21)